Shelly - move coordinators to coordinator.py
